# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaSForge API",
        "description": "OpenAPI documentation for the SaaSForge multi-tenant API.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current application host"
        }
    ],
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List projects in the current workspace",
                "operationId": "projectsIndex",
                "parameters": [
                    {
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "description": "Workspace context header (one resolution strategy)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Comma-separated statuses",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "active,archived"
                        }
                    },
                    {
                        "name": "created_by",
                        "in": "query",
                        "description": "User ID or 'me'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "me"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Comma-separated sortable fields, prefix with - for desc",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "-created_at,name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project collection"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Workspace not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a project in the current workspace",
                "operationId": "projectsStore",
                "parameters": [
                    {
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "description": "Workspace context header (one resolution strategy)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Marketing Website"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Workspace not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/projects/{project}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Show a project",
                "operationId": "projectsShow",
                "parameters": [
                    {
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "description": "Workspace context header (one resolution strategy)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Soft-delete a project",
                "operationId": "projectsDestroy",
                "parameters": [
                    {
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "description": "Workspace context header (one resolution strategy)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update a project",
                "operationId": "projectsUpdate",
                "parameters": [
                    {
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "description": "Workspace context header (one resolution strategy)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "archived"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/projects/{id}/restore": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Restore a soft-deleted project",
                "operationId": "projectsRestore",
                "parameters": [
                    {
                        "name": "X-Workspace-ID",
                        "in": "header",
                        "description": "Workspace context header (one resolution strategy)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project restored"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Paste a Sanctum bearer token in the Authorization header.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}